modularize cropAndResize
