+++ tfjs-backend-wasm/src/kernels/CropAndResize.ts
@@ -15,22 +15,11 @@
  * =============================================================================
  */
 
-import {NamedAttrMap, NamedTensorInfoMap, registerKernel, TensorInfo} from '@tensorflow/tfjs-core';
+import {CropAndResize, CropAndResizeAttrs, CropAndResizeInputs, NamedAttrMap, NamedTensorInfoMap, registerKernel, TensorInfo} from '@tensorflow/tfjs-core';
 
 import {BackendWasm} from '../backend_wasm';
+
 import {cast} from './Cast';
-
-interface CropAndResizeInputs extends NamedTensorInfoMap {
-  images: TensorInfo;
-  boxes: TensorInfo;
-  boxInd: TensorInfo;
-}
-
-interface CropAndResizeAttrs extends NamedAttrMap {
-  method: keyof InterpolationMethod;
-  extrapolationValue: number;
-  cropSize: [number, number];
-}
 
 // Must match enum in CropAndResize.cc
 enum InterpolationMethod {
@@ -60,23 +49,23 @@ function setup(backend: BackendWasm): void {
 
 function cropAndResize(args: {
   backend: BackendWasm,
-  inputs: CropAndResizeInputs,
-  attrs: CropAndResizeAttrs
+  inputs: NamedTensorInfoMap,
+  attrs: NamedAttrMap
 }): TensorInfo {
   const {backend, inputs, attrs} = args;
-  const {method, extrapolationValue, cropSize} = attrs;
-  const {images, boxes, boxInd} = inputs;
+  const {method, extrapolationValue, cropSize} =
+      attrs as {} as CropAndResizeAttrs;
+  const {image, boxes, boxInd} = inputs as CropAndResizeInputs;
 
   const numBoxes = boxes.shape[0];
 
   const [cropHeight, cropWidth] = cropSize as [number, number];
-  const outShape = [numBoxes, cropHeight, cropWidth, images.shape[3]];
+  const outShape = [numBoxes, cropHeight, cropWidth, image.shape[3]];
 
-  let imagesData = backend.dataIdMap.get(images.dataId);
+  let imagesData = backend.dataIdMap.get(image.dataId);
   let castedData;
-  if (images.dtype !== 'float32') {
-    castedData =
-        cast({backend, inputs: {x: images}, attrs: {dtype: 'float32'}});
+  if (image.dtype !== 'float32') {
+    castedData = cast({backend, inputs: {x: image}, attrs: {dtype: 'float32'}});
     imagesData = backend.dataIdMap.get(castedData.dataId);
   }
 
@@ -87,7 +76,7 @@ function cropAndResize(args: {
   const out = backend.makeOutput(outShape, 'float32');
   const outId = backend.dataIdMap.get(out.dataId).id;
 
-  const imagesShapeBytes = new Uint8Array(new Int32Array(images.shape).buffer);
+  const imagesShapeBytes = new Uint8Array(new Int32Array(image.shape).buffer);
 
   wasmCropAndResize(
       imagesId, boxesId, boxIndId, numBoxes, imagesShapeBytes, cropHeight,
@@ -103,7 +92,7 @@ function cropAndResize(args: {
 }
 
 registerKernel({
-  kernelName: 'CropAndResize',
+  kernelName: CropAndResize,
   backendName: 'wasm',
   setupFunc: setup,
   kernelFunc: cropAndResize
