+++ tfjs-core/src/tests.ts
@@ -66,6 +66,7 @@ import './ops/conv2d_transpose_test';
 import './ops/conv3d_test';
 import './ops/conv3d_transpose_test';
 import './ops/conv_util_test';
+import './ops/crop_and_resize_test';
 import './ops/cumsum_test';
 import './ops/depth_to_space_test';
 import './ops/diag_test';
